+++ Tests/INCommonsTests/Swift/Types/ReleaseTrigger/ReleaseTriggerTests.swift
@@ -1,0 +1,17 @@
+import INCommons
+import XCTest
+
+final class ReleaseTriggerTests: XCTestCase {
+	func testClosureGetsCalledOnDeinit() {
+		let callbackExpectation = expectation(description: "callbackExpectation")
+		var sut: ReleaseTrigger<Int>? = ReleaseTrigger<Int>(reference: 3) { reference in
+			XCTAssertEqual(reference, 3)
+			callbackExpectation.fulfill()
+		}
+		weak var weakSut = sut
+		XCTAssertNotNil(weakSut)
+		sut = nil
+		XCTAssertNil(weakSut)
+		waitForExpectations(timeout: 1)
+	}
+}